Moved files to allow npm run build to work
